test: fix test failures and move mock to __mocks__ directory
- Update test to match new message display behavior
- Move mockGetStepInfo to __mocks__ directory to avoid Jest treating it as a test file
