+++ src/utils/__mocks__/mockGetStepInfo.ts
